requirements.txt
# Core dependencies for Phase 1
torch>=2.0.0
transformers>=4.30.0
Pillow>=10.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
pytest>=7.4.0
japanese-clip @ git+https://github.com/tattn/japanese-clip.git
